+++ CHANGELOG.md
@@ -7,4 +7,5 @@
 
 ### [teleport_warmup] 
 - feature: now the admin players can bypass the teleport warmup.
+- feature: now the entity with `age <= 3` can bypass the teleport warmup. (Making it friendly for `a teleport after respawn`)